move check_for_zmq to utils.zmqrelated
